feat: add commonjs plugin to handle require
core-js (unfortunately) makes use of require, so we need the commonjs
plugin so that rollup sees an import realises it has to fetch the source
and bundles it.  It's very clever, though, if you stop using whatever
core-js was polyfilling for you then rollup doesn't just drop that it
drops the commonjs transform, too.